refactors the file structure
