+++ client/src/authentication/Login.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Redirect} from 'react-router-dom';
 
 
@@ -7,6 +7,10 @@ function Login(props)  {
       username: '',
       password: ''
   });
+
+  useEffect(() => {
+    console.log('LOGIN props: ', props)
+  }, [])
 
   const handleChange = e => {
       setUser({
