+++ client/src/components/Home.js
@@ -1,23 +1,18 @@
 import React, { useEffect } from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
-
+import NavBar from './header/NavBar';
+import Panel from './main/Panel';
 
 function Home(props) {
 
   useEffect(() => {
-
+    console.log('HOME props: ', props)
   }, [])
-
-  const logout = () => {
-    props.logOut(props.history);
-  }
   
   return (
     <div>
-      <div>
-        <h1>Home Page</h1>
-        <button onClick={logout}>Logout</button>
-      </div>
+      <NavBar {...props}/>
+      <Panel {...props}/>
     </div>
   ) 
 }
