+++ client/src/components/main/Panel.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Panel() {
+
+  return (
+    <div>
+      Panel
+    </div>
+  )
+}
+
+export default Panel;
